+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file
 ## [1.12.5] -- 2020-4-15
 ### Changed
 - `music.Song.fillTrack()` uses separate channels for instruments
+- `music.Song.fillTrack()` no longer creates MIDI events for rests
 
 ## [1.12.4] -- 2020-4-14
 ### Added